SONARPHP-1520 Address comments
