+++ its/plugin/projects/custom_rules/src/simple.php
@@ -1,11 +1,6 @@
 <?php
 
-function f() {
-
-  $f1 = function () {  // NOK - function expression
-  };
-
-  for (;;) {           // NOK - for statement
-  }
-
-}
+some_function();
+foo();
+test();
+buzz();
