+++ php-custom-rules-plugin/src/main/java/org/sonar/samples/php/PHPCustomRulesPlugin.java
