+++ php-custom-rules-plugin/src/main/java/org/sonar/samples/php/checks/ForbiddenFunctionUseCheck.java
@@ -19,54 +19,50 @@
  */
 package org.sonar.samples.php.checks;
 
-import java.util.List;
 import java.util.Set;
 import org.sonar.check.Rule;
-import org.sonar.plugins.php.api.tree.Tree;
 import org.sonar.plugins.php.api.tree.Tree.Kind;
 import org.sonar.plugins.php.api.tree.declaration.NamespaceNameTree;
 import org.sonar.plugins.php.api.tree.expression.ExpressionTree;
 import org.sonar.plugins.php.api.tree.expression.FunctionCallTree;
-import org.sonar.plugins.php.api.visitors.PHPSubscriptionCheck;
 import org.sonar.plugins.php.api.visitors.PHPVisitorCheck;
 
 /**
- * Example of implementation of a check by extending {@link PHPSubscriptionCheck}.
- * PHPSubscriptionCheck provides method {@link PHPSubscriptionCheck#visitNode} to visit nodes of the Abstract Syntax Tree
- * that represents the source code. Provide the list of nodes you want to visit through {@link PHPSubscriptionCheck#nodesToVisit}.
+ * Example of implementation of a check by extending {@link PHPVisitorCheck}.
+ * PHPVisitorCheck provides methods to visit nodes of the Abstract Syntax Tree
+ * that represents the source code.
  * <p>
- * Those methods should be overridden to process information
+ * Those methods can be overridden to process information
  * related to node and issue can be created via the context that can be
  * accessed through {@link PHPVisitorCheck#context()}.
  */
 @Rule(
-  key = OtherForbiddenFunctionUseCheck.KEY,
+  key = ForbiddenFunctionUseCheck.KEY,
   name = "Forbidden function should not be used.",
-  tags = {"convention"},
-  // Description can either be given in this annotation or through HTML name <ruleKey>.html located in package
-  // src/resources/org/sonar/l10n/php/rules/<repositoryKey>
-  description = "<p>The following functions should not be used:</p> <ul><li>foo</li> <li>bar</li></ul>")
-public class OtherForbiddenFunctionUseCheck extends PHPSubscriptionCheck {
+  tags = {"convention"}
+// Description can either be given in this annotation or through HTML name <ruleKey>.html located in package
+// src/resources/org/sonar/l10n/php/rules/<repositoryKey>
+// description = "<p>The following functions should not be used:</p> <ul><li>foo</li> <li>bar</li></ul>"
+)
+public class ForbiddenFunctionUseCheck extends PHPVisitorCheck {
 
+  public static final String KEY = "S1";
   private static final Set<String> FORBIDDEN_FUNCTIONS = Set.of("foo", "bar");
-  public static final String KEY = "S2";
-
-  @Override
-  public List<Kind> nodesToVisit() {
-    return List.of(Kind.FUNCTION_CALL);
-  }
 
   /**
    * Overriding method visiting the call expression to create an issue
    * each time a call to "foo()" or "bar()" is done.
    */
   @Override
-  public void visitNode(Tree tree) {
-    ExpressionTree callee = ((FunctionCallTree) tree).callee();
+  public void visitFunctionCall(FunctionCallTree tree) {
+    ExpressionTree callee = tree.callee();
 
     if (callee.is(Kind.NAMESPACE_NAME) && FORBIDDEN_FUNCTIONS.contains(((NamespaceNameTree) callee).qualifiedName())) {
       context().newIssue(this, callee, "Remove the usage of this forbidden function.");
     }
+
+    // super method must be called in order to visit function call node's children
+    super.visitFunctionCall(tree);
   }
 
 }
