+++ php-custom-rules-plugin/src/main/resources/org/sonar/l10n/php/rules/custom/S1.html
