+++ php-custom-rules-plugin/src/main/resources/org/sonar/l10n/php/rules/custom/S2.html
@@ -2,6 +2,6 @@
     The following functions should not be used:
 </p>
 <ul>
-    <li>foo</li>
-    <li>bar</li>
+    <li>fizz</li>
+    <li>buzz</li>
 </ul>
