+++ php-custom-rules-plugin/src/test/java/org/sonar/samples/php/MyPhpRulesTest.java
