+++ php-custom-rules-plugin/src/test/resources/checks/forbiddenFunctionUseCheck.php
