+++ settings.gradle.kts
@@ -34,9 +34,8 @@ includeBuild("build-logic")
 include(":sonar-php-plugin")
 include(":php-frontend")
 include(":php-checks")
-include(":php-custom-rules")
+include(":php-custom-rules-plugin")
 include(":its:plugin:tests")
-include(":its:plugin:plugins:php-custom-rules-plugin")
 include(":its:ruling")
 
 gradle.allprojects {
